% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{get_raster_vals}
\alias{get_raster_vals}
\title{Extract value of rasters at specified lat/long points}
\usage{
get_raster_vals(
  conn = repel_remote_conn(),
  raster_name,
  lon,
  lat,
  check = TRUE
)
}
\arguments{
\item{conn}{Connection to the REPEL database}

\item{raster_name}{Name of a raster table, typically starting with `raster_`}

\item{lon}{Longitude, between -180 and 180}

\item{lat}{Latitude, between -90 and 90}

\item{check}{Whether to fail early with incorrect bounds or table names}
}
\value{
A data frame with lat/lon and raster values
}
\description{
Extract value of rasters at specified lat/long points
}
