% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{repel_fields_schema}
\alias{repel_fields_schema}
\title{View field descriptions for REPEL data}
\usage{
repel_fields_schema()
}
\value{
a tibble object
}
\description{
View field descriptions for REPEL data
}
