% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect-local.R
\name{repel_local_conn}
\alias{repel_local_conn}
\title{The local REPEL database}
\usage{
repel_local_conn(dbdir = repel_local_path())
}
\arguments{
\item{dbdir}{The location of the database on disk. Defaults to
`repeldata` under [rappdirs::user_data_dir()], or the environment variable `WAHIS_DB_DIR`.}
}
\value{
A database connection
}
\description{
Returns a connection to the local REPEL database. This is a DBI-compliant
[duckdb::duckdb()] database connection.
}
