% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onattach.R
\name{repel_local_delete}
\alias{repel_local_delete}
\title{Remove the local WAHIS database}
\usage{
repel_local_delete()
}
\description{
Deletes all tables from the local database
}
\examples{
\donttest{
\dontrun{
repel_local_delete()
}
}
}
