% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect-local.R
\name{repel_local_disconnect}
\alias{repel_local_disconnect}
\title{Disconnect from the WAHIS database}
\usage{
repel_local_disconnect()
}
\description{
A utility function for disconnecting from the database.
}
\examples{
repel_local_disconnect()
}
