% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection-pane.R
\name{repel_local_pane}
\alias{repel_local_pane}
\title{Show REPEL (local) database in the RStudio Connections Pane}
\usage{
repel_local_pane()
}
\value{

}
\description{
Show REPEL (local) database in the RStudio Connections Pane
}
