% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection-pane.R
\name{repel_pane}
\alias{repel_pane}
\title{Open WAHIS database connection pane in RStudio}
\usage{
repel_pane(
  conn = repel_local_conn(),
  connectCode = "repeldata::repel_local_pane()",
  actions = list()
)
}
\description{
This function launches the RStudio "Connection" pane to interactively
explore the database.
}
\examples{
if (!is.null(getOption("connectionObserver"))) wahis_pane()

}
