% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect-remote.R
\name{repel_remote_conn}
\alias{repel_remote_conn}
\title{Connect to the remote REPEL database}
\usage{
repel_remote_conn(host = NULL, port = NULL, user = NULL, password = NULL)
}
\arguments{
\item{host}{The host URL.  Defaults to the EHA server}

\item{port}{the host connection port.  Generally does not need to be set}

\item{user}{User login name.  In general to be set in `.Renviron`}

\item{password}{User login passwword. In general to be set in `.Renviron`}
}
\value{
a DBI connection object
}
\description{
Returns a connection to the remote REPEL database, hosted on a Postgres server.
`REPELDATA_USER` and `REPELDATA_PASSWORD` should be set in your environment
(e.g. `.Renviron` file) to provide access to the data.
}
