% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect-remote.R
\name{repel_remote_disconnect}
\alias{repel_remote_disconnect}
\title{Disconnect from the remote REPEL database}
\usage{
repel_remote_disconnect()
}
\value{

}
\description{
Disconnect from the remote REPEL database
}
