% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{repel_remote_pane}
\alias{repel_remote_pane}
\title{Show REPEL (remote) database in the RStudio Connections Pane}
\usage{
repel_remote_pane()
}
\value{

}
\description{
Show REPEL (remote) database in the RStudio Connections Pane
}
