% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{repel_schema}
\alias{repel_schema}
\title{View field descriptions for REPEL data
Database documention is available here: https://github.com/ecohealthalliance/repel-docs/blob/master/database-schema.md}
\usage{
repel_schema()
}
\value{
a tibble object
}
\description{
View field descriptions for REPEL data
Database documention is available here: https://github.com/ecohealthalliance/repel-docs/blob/master/database-schema.md
}
