% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection-pane.R
\name{wahis_pane}
\alias{wahis_pane}
\title{Open WAHIS database connection pane in RStudio}
\usage{
wahis_pane(
  conn = wahis_db(),
  dbname = "REPEL local database",
  connectCode = "wahisclient::wahis_pane()",
  disconnect = wahisclient::wahis_disconnect
)
}
\description{
This function launches the RStudio "Connection" pane to interactively
explore the database.
}
\examples{
if (!is.null(getOption("connectionObserver"))) wahis_pane()
}
