% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{repel_local_conn}
\alias{repel_local_conn}
\title{The local WAHIS database}
\usage{
repel_local_conn(dbdir = repel_local_path())
}
\arguments{
\item{dbdir}{The location of the database on disk. Defaults to
`wahisclient` under [rappdirs::user_data_dir()], or the environment variable `WAHIS_DB_DIR`.}
}
\value{
A MonetDBLite DBI connection
}
\description{
Returns a connection to the local WAHIS database. This is a DBI-compliant
[MonetDBLite::MonetDBLite()] database connection. When using **dplyr**-based
workflows, one typically accesses tables with functions such as
[wahis_shipments()], but this function lets one interact with the database
directly via SQL.
}
